HTTP Request to RAE website: Cloudflare problem
RAE website have some sort of DDOS protection using Cloudflare anti-spam
systems. When requesting a HTML page from RAE, cloudflare intercepts it
and return their own HTML page.

I've tried to add a User-Agent HTTP header for Cloudflare to think is a
web browser, but problem is still happening. Need to work on this for a
solution.